world/navigation/requirements.txt
redis>=5.2.1
azure-messaging-webpubsubservice>=1.2.1
azure-messaging-webpubsubclient>=1.1.0
azure-identity>=1.19.0
